IPCweights {mboost}	R Documentation

Inverse Probability of Censoring Weights

Description

Compute weights for censored regression via the inverted probability of censoring principle.

Usage

IPCweights(x, maxweight = 5)

Arguments

x
an object of class Surv.

maxweight
the maximal value of the returned weights.

Details

Inverse probability of censoring weights are one possibility to fit models formulated in the full data world in the presence of censoring, i.e., the observed data world, see van der Laan and Robins (2003) for the underlying theory and Hothorn et al. (2006) for an application to survival analysis.

Value

A vector of numeric weights.

References

Mark J. van der Laan and James M. Robins (2003), Unified Methods for Censored Longitudinal Data and Causality, Springer, New York.

Torsten Hothorn, Peter Buehlmann, Sandrine Dudoit, Annette Molinaro and Mark J. van der Laan (2006), Survival ensembles. Biostatistics 7(3), 355–373.

Peter Buehlmann and Torsten Hothorn (2007), Boosting algorithms: regularization, prediction and model fitting. Statistical Science, 22(4), 477–505.

[Package mboost version 2.9-11 Index]