boost_family-class {mboost}	R Documentation

Class "boost\_family": Gradient Boosting Family

Description

Objects of class boost_family define negative gradients of loss functions to be optimized.

Objects from the Class

Objects can be created by calls of the form Family(...)

Slots

ngradient:

a function with arguments y and f implementing the negative gradient of the loss function.

risk:

a risk function with arguments y, f and w, the weighted mean of the loss function by default.

offset:

a function with argument y and w (weights) for computing a scalar offset.

weights:

a logical indicating if weights are allowed.

check_y:

a function for checking the class / mode of a response variable.

nuisance:

a function for extracting nuisance parameters.

response:

inverse link function of a GLM or any other transformation on the scale of the response.

rclass:

function to derive class predictions from conditional class probabilities (for models with factor response variable).

name:

a character giving the name of the loss function for pretty printing.

charloss:

a character, the deparsed loss function.

See Also

Family

Examples

    Laplace()

[Package mboost version 2.6-0 Index]